backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
databases
pydantic
passlib[bcrypt]
python-multipart
pytest
httpx
alembic
python-dotenv
corsheaders